refactor: Garage, RaceCount 클래스 적용 및 테스트 코드 작성
변경 전: RacingGameManger가 사용자 입력을 모두 검증함
변경 후: RacingGameManager가 Garage, RaceCount로 분리하도록 변경
변경 이유: RacingGameManager가 너무 많은 책임을 가져서 기능을 분리
RacingGameManager는 Garage와 RaceCount로 분리되었음으로 미사용하기 때문에 제거한다.
