+++ src/main/java/racingcar/config/RacingGameConfig.java
@@ -1,17 +1,12 @@
 package racingcar.config;
 
 import racingcar.util.GameResultMaker;
-import racingcar.util.RacingGameManager;
 import racingcar.controller.RacingGame;
-import racingcar.view.OutputView;
-
-import java.util.ArrayList;
 
 public class RacingGameConfig {
 
     public RacingGame initRacingGame() {
-        RacingGameManager racingGameManager = new RacingGameManager(new ArrayList<>());
         GameResultMaker gameResultMaker = new GameResultMaker();
-        return new RacingGame(racingGameManager, gameResultMaker);
+        return new RacingGame(gameResultMaker);
     }
 }
